adds conditional filtering to single filtering file
