+++ q2_feature_table/__init__.py
@@ -11,12 +11,12 @@ from ._subsample import subsample
 from ._transform import (presence_absence, relative_frequency, transpose)
 from ._summarize import (summarize, tabulate_seqs)
 from ._merge import (merge, merge_seqs, merge_taxa, overlap_methods)
-from ._filter import (filter_samples, filter_features, filter_seqs)
+from ._filter import (filter_samples, filter_features, filter_seqs, 
+					  filter_features_conditionally)
 from ._core_features import core_features
 from ._group import group
 from ._rename import rename_ids
 from ._heatmap import (heatmap, heatmap_choices)
-from ._conditional_filter import (filter_features_conditionally)
 from ._version import get_versions
 
 __version__ = get_versions()['version']
